feat(ui): separate KVM table into LXD and virsh KVMs
