+++ ui/src/app/kvm/views/KVMList/CPUColumn/CPUColumn.tsx
