+++ ui/src/app/kvm/views/KVMList/CPUColumn/CPUPopover/CPUPopover.test.tsx
