+++ ui/src/app/kvm/views/KVMList/CPUColumn/CPUPopover/_index.scss
