+++ ui/src/app/kvm/views/KVMList/CPUColumn/CPUPopover/index.ts
