+++ ui/src/app/kvm/views/KVMList/CPUColumn/index.ts
