--- ui/src/app/kvm/views/KVMList/KVMListTable/OSColumn/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./OSColumn";
