--- ui/src/app/kvm/views/KVMList/KVMListTable/PowerColumn/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./PowerColumn";
