--- ui/src/app/kvm/views/KVMList/KVMListTable/TypeColumn/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TypeColumn";
