--- ui/src/app/kvm/views/KVMList/KVMListTable/_index.scss
@@ -1,51 +1,0 @@
-@mixin KVMListTable {
-  .kvm-list-table {
-    td,
-    th {
-      // FQDN
-      &:nth-child(1) {
-        @include breakpoint-widths(60%, 60%, 60%, 24%, 24%);
-      }
-
-      // Power
-      &:nth-child(2) {
-        @include breakpoint-widths(0, 0, 0, 19%, 15%);
-      }
-
-      // KVM host type
-      &:nth-child(3) {
-        @include breakpoint-widths(40%, 40%, 40%, 18%, 15%);
-      }
-
-      // VMs / Owners
-      &:nth-child(4) {
-        @include breakpoint-widths(0, 0, 0, 0, 9%);
-      }
-
-      // OS
-      &:nth-child(5) {
-        @include breakpoint-widths(0, 0, 0, 0, 21%);
-      }
-
-      // Resource pool / AZ
-      &:nth-child(6) {
-        @include breakpoint-widths(0, 0, 0, 23%, 16%);
-      }
-
-      // CPU
-      &:nth-child(7) {
-        @include breakpoint-widths(0, 5rem, 9rem, 9rem, 9rem);
-      }
-
-      // RAM
-      &:nth-child(8) {
-        @include breakpoint-widths(0, 8rem, 11rem, 11rem, 11rem);
-      }
-
-      // Storage
-      &:nth-child(9) {
-        @include breakpoint-widths(0, 8rem, 11rem, 11rem, 11rem);
-      }
-    }
-  }
-}
