+++ ui/src/app/kvm/views/KVMList/LxdTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./LxdTable";
