+++ ui/src/app/kvm/views/KVMList/NameColumn/NameColumn.test.tsx
