+++ ui/src/app/kvm/views/KVMList/NameColumn/index.ts
