+++ ui/src/app/kvm/views/KVMList/RAMColumn/RAMColumn.test.tsx
