+++ ui/src/app/kvm/views/KVMList/RAMColumn/RAMColumn.tsx
