+++ ui/src/app/kvm/views/KVMList/RAMColumn/RAMPopover/index.ts
