+++ ui/src/app/kvm/views/KVMList/RAMColumn/index.ts
