+++ ui/src/app/kvm/views/KVMList/VirshTable/VirshTable.tsx
@@ -1,62 +1,26 @@
-import { useEffect } from "react";
-
 import { Col, MainTable, Row } from "@canonical/react-components";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 
-import CPUColumn from "./CPUColumn";
-import NameColumn from "./NameColumn";
-import OSColumn from "./OSColumn";
-import PoolColumn from "./PoolColumn";
-import PowerColumn from "./PowerColumn";
-import RAMColumn from "./RAMColumn";
-import StorageColumn from "./StorageColumn";
-import TypeColumn from "./TypeColumn";
-import VMsColumn from "./VMsColumn";
+import CPUColumn from "../CPUColumn";
+import NameColumn from "../NameColumn";
+import PoolColumn from "../PoolColumn";
+import RAMColumn from "../RAMColumn";
+import StorageColumn from "../StorageColumn";
+import VMsColumn from "../VMsColumn";
 
-import { general as generalActions } from "app/base/actions";
 import TableHeader from "app/base/components/TableHeader";
 import { useTableSort } from "app/base/hooks";
-import type { TSFixMe } from "app/base/types";
-import { actions as controllerActions } from "app/store/controller";
-import type { Controller } from "app/store/controller/types";
-import generalSelectors from "app/store/general/selectors";
-import { actions as machineActions } from "app/store/machine";
-import type { Machine } from "app/store/machine/types";
-import { actions as podActions } from "app/store/pod";
 import podSelectors from "app/store/pod/selectors";
 import type { Pod } from "app/store/pod/types";
-import { actions as poolActions } from "app/store/resourcepool";
 import poolSelectors from "app/store/resourcepool/selectors";
 import type { ResourcePool } from "app/store/resourcepool/types";
-import { actions as zoneActions } from "app/store/zone";
-import { getStatusText } from "app/utils";
 
-type SortKey = keyof Pod | "cpu" | "os" | "pool" | "power" | "ram" | "storage";
+type SortKey = keyof Pod | "cpu" | "pool" | "ram" | "storage";
 
-const getSortValue = (
-  sortKey: SortKey,
-  kvm: Pod,
-  kvmHosts: (Controller | Machine)[],
-  pools: ResourcePool[],
-  osReleases: TSFixMe
-) => {
-  const kvmHost = kvmHosts.find((host) => host.system_id === kvm.host);
+const getSortValue = (sortKey: SortKey, kvm: Pod, pools: ResourcePool[]) => {
   const kvmPool = pools.find((pool) => kvm.pool === pool.id);
 
   switch (sortKey) {
-    case "power":
-      if (kvmHost && "power_state" in kvmHost) {
-        return kvmHost.power_state;
-      }
-      return "unknown";
-    case "os":
-      if (kvmHost && kvmHost.osystem in osReleases) {
-        return getStatusText(
-          kvmHost,
-          `${kvmHost.osystem}/${kvmHost.distro_series}`
-        );
-      }
-      return "unknown";
     case "pool":
       return kvmPool?.name || "unknown";
     case "cpu":
@@ -75,10 +39,7 @@ const generateRows = (kvms: Pod[]) =>
     key: kvm.id,
     columns: [
       { content: <NameColumn id={kvm.id} /> },
-      { content: <PowerColumn id={kvm.id} /> },
-      { content: <TypeColumn id={kvm.id} /> },
       { className: "u-align--right", content: <VMsColumn id={kvm.id} /> },
-      { content: <OSColumn id={kvm.id} /> },
       { content: <PoolColumn id={kvm.id} /> },
       { content: <CPUColumn id={kvm.id} /> },
       { content: <RAMColumn id={kvm.id} /> },
@@ -86,13 +47,9 @@ const generateRows = (kvms: Pod[]) =>
     ],
   }));
 
-const KVMListTable = (): JSX.Element => {
-  const dispatch = useDispatch();
-  const osReleases = useSelector(generalSelectors.osInfo.getAllOsReleases);
-  const kvms = useSelector(podSelectors.kvms);
-  const kvmHosts = useSelector(podSelectors.getAllHosts);
+const VirshTable = (): JSX.Element => {
+  const virshKvms = useSelector(podSelectors.virsh);
   const pools = useSelector(poolSelectors.all);
-
   const { currentSort, sortRows, updateSort } = useTableSort<Pod, SortKey>(
     getSortValue,
     {
@@ -100,17 +57,7 @@ const KVMListTable = (): JSX.Element => {
       direction: "descending",
     }
   );
-
-  useEffect(() => {
-    dispatch(controllerActions.fetch());
-    dispatch(generalActions.fetchOsInfo());
-    dispatch(machineActions.fetch());
-    dispatch(podActions.fetch());
-    dispatch(poolActions.fetch());
-    dispatch(zoneActions.fetch());
-  }, [dispatch]);
-
-  const sortedKVMs = sortRows(kvms, kvmHosts, pools, osReleases);
+  const sortedKVMs = sortRows(virshKvms, pools);
 
   return (
     <Row>
@@ -131,31 +78,6 @@ const KVMListTable = (): JSX.Element => {
               ),
             },
             {
-              content: (
-                <TableHeader
-                  className="p-double-row__header-spacer"
-                  currentSort={currentSort}
-                  data-test="power-header"
-                  onClick={() => updateSort("power")}
-                  sortKey="power"
-                >
-                  Power
-                </TableHeader>
-              ),
-            },
-            {
-              content: (
-                <TableHeader
-                  currentSort={currentSort}
-                  data-test="type-header"
-                  onClick={() => updateSort("type")}
-                  sortKey="type"
-                >
-                  KVM host type
-                </TableHeader>
-              ),
-            },
-            {
               className: "u-align--right",
               content: (
                 <>
@@ -169,19 +91,6 @@ const KVMListTable = (): JSX.Element => {
                   </TableHeader>
                   <TableHeader>Owners</TableHeader>
                 </>
-              ),
-            },
-            {
-              content: (
-                <TableHeader
-                  className="p-double-row__header-spacer"
-                  currentSort={currentSort}
-                  data-test="os-header"
-                  onClick={() => updateSort("os")}
-                  sortKey="os"
-                >
-                  OS
-                </TableHeader>
               ),
             },
             {
@@ -244,4 +153,4 @@ const KVMListTable = (): JSX.Element => {
   );
 };
 
-export default KVMListTable;
+export default VirshTable;
