+++ ui/src/app/kvm/views/KVMList/VirshTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./VirshTable";
